requirements.txt
streamlit
Pillow
pinecone-client
python-dotenv
groq
